requirements.txt
asgiref==3.5.2
dj-database-url==0.5.0
Django==4.0.5
django-bootstrap-form==3.4
django-csp==3.7
django-heroku==0.3.1
django-widget-tweaks==1.4.12
gunicorn==20.1.0
mysqlclient==2.1.0
mysql-connector-python=8.0.29
psycopg2==2.9.3
PyMySQL==1.0.2
sqlparse==0.4.2
tzdata==2022.1
upgrade-requirements==1.7.0
whitenoise==6.2.0
